Add test/fixtures folder and testDB.js. Modify bamazonDB.test.js.
Reorganize sql files for creation of test database. Add testDB.js
module which provides methods to initialize a test database and to
seed the database either from file or from agruments containing the
values to seed the database with.

Modify bamazonDB.test.js module to implement the new testDB fixtures.
Skip testing methods which must be updated for the addition of the
foreign key in the products table.